+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 * Updated documentation
 * Changed export order so haddock makes better documentation
 * Added tested-with in cabal file
+* No longer exports vkvParser